float 32 input since voc inputs are going to be in float32 datatype

+++ optimize.py
@@ -28,7 +28,7 @@ converter.representative_dataset = representative_dataset_gen
 
 converter.target_spec.supported_ops = [tf.lite.OpsSet.TFLITE_BUILTINS_INT8]
 
-converter.inference_input_type = tf.uint8
+converter.inference_input_type = tf.float32
 converter.inference_output_type = tf.uint8
 
 tf_lite_model = converter.convert()
